+++ CHANGELOG.md
@@ -8,6 +8,7 @@ WARNING!: This version have so many break changes!
 - Use C ext instead of ImageMagick, now it's no dependencies!
 - New captcha style.
 - Remove `len`, `font_size`, `cache_limit` config key, no support now.
+- Output `GIF` format.
 
 1.2.0
 -----